ensure consistency in update-version-pinning.sh
Prior to this commit we ran the risk that `nix-prefetch-url` would reach for a different nixpkgs.tar.gz file than we got from github via wget.

I solve that by removing the wget command and changing the `nix-prefetch-url` command to _not_ unpack the tar file.
Then we just hash the file nix grabbed for us directly.