Revert "feat(default.nix): Rust to beta for RFC feature(trait-upcasting)"
This reverts commit 08722ab2fb04beb5c023027d610111e910dbc890.

This commit also bumps the version for stable to 1.86.0.

Now that Rust v1.86.0 is out with the features we wanted for the
dataplane, let's use the stable toolchain as a default again.

Note: I dropped the Nix versions hash changes, I think they were
unrelated to the change of toolchain and I don't know how to refresh
them anyway. I'll leave it to the GitHub workflow.

Signed-off-by: Quentin Monnet <qmo@qmon.net>