fix(frr): rework frr build
The previous FRR build relied on an unsound hack to make sure the frr agent's modules ended up in the right location in
the container. This is silly.

Basically, we need to build a normal FHS style directory structure and then toss that in /nix rather than trying to foce
FRR into a shape that makes no sense to it using rpath and shebang tricks.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>