fix(compile-env): add openssl to compile env
Unfortunately, some cargo dependencies are linked against openssl instead
of rustls.

Enabling openssl in the compile-env allows more cargo-install'ed tools to
function without openssl.

Ideally we would be able to exhaustively list all of the cargo tooling we
need and just static link it to whatever TLS provider it needs, but that
strategy prevents the dev from using `cargo install` for other tools
they may like but which are not used in CI (e.g. LSP servers for cargo).

Signed-off-by: Daniel Noland <daniel@githedgehog.com>