chore: syntax and structure cleanup
This commit does two things to clean up the formatting of the main
default.nix file.

First, it factors the optimizedBuild function out, mostly for consistency.
This does not actually impact the meaning of the expressions.

Second, it removes a (now) pointless split of the libyang build into a -dynamic,
-static, and a combined build.
The -dynamic build is no longer useful now that our build of FRR is almost
completely static, so this split was just pointless complexity.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>